Calculates ace based transformations and correlation, handling missing values and factors.

Calculates ace based transformations and correlation, handling missing values and factors.

Usage

ace_cor(x, y, handle.na = TRUE)

Arguments

x

a numeric vector or factor

y

a numeric vector or factor

handle.na

If TRUE uses pairwise complete observations.

Value

result of acepack::ace

Examples

ace_cor(iris$Sepal.Length, iris$Species)
#> $p
#> [1] 1
#> 
#> $n
#> [1] 150
#> 
#> $x
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,]  5.1  4.9  4.7  4.6    5  5.4  4.6    5  4.4   4.9   5.4   4.8   4.8   4.3
#>      [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25] [,26]
#> [1,]   5.8   5.7   5.4   5.1   5.7   5.1   5.4   5.1   4.6   5.1   4.8     5
#>      [,27] [,28] [,29] [,30] [,31] [,32] [,33] [,34] [,35] [,36] [,37] [,38]
#> [1,]     5   5.2   5.2   4.7   4.8   5.4   5.2   5.5   4.9     5   5.5   4.9
#>      [,39] [,40] [,41] [,42] [,43] [,44] [,45] [,46] [,47] [,48] [,49] [,50]
#> [1,]   4.4   5.1     5   4.5   4.4     5   5.1   4.8   5.1   4.6   5.3     5
#>      [,51] [,52] [,53] [,54] [,55] [,56] [,57] [,58] [,59] [,60] [,61] [,62]
#> [1,]     7   6.4   6.9   5.5   6.5   5.7   6.3   4.9   6.6   5.2     5   5.9
#>      [,63] [,64] [,65] [,66] [,67] [,68] [,69] [,70] [,71] [,72] [,73] [,74]
#> [1,]     6   6.1   5.6   6.7   5.6   5.8   6.2   5.6   5.9   6.1   6.3   6.1
#>      [,75] [,76] [,77] [,78] [,79] [,80] [,81] [,82] [,83] [,84] [,85] [,86]
#> [1,]   6.4   6.6   6.8   6.7     6   5.7   5.5   5.5   5.8     6   5.4     6
#>      [,87] [,88] [,89] [,90] [,91] [,92] [,93] [,94] [,95] [,96] [,97] [,98]
#> [1,]   6.7   6.3   5.6   5.5   5.5   6.1   5.8     5   5.6   5.7   5.7   6.2
#>      [,99] [,100] [,101] [,102] [,103] [,104] [,105] [,106] [,107] [,108]
#> [1,]   5.1    5.7    6.3    5.8    7.1    6.3    6.5    7.6    4.9    7.3
#>      [,109] [,110] [,111] [,112] [,113] [,114] [,115] [,116] [,117] [,118]
#> [1,]    6.7    7.2    6.5    6.4    6.8    5.7    5.8    6.4    6.5    7.7
#>      [,119] [,120] [,121] [,122] [,123] [,124] [,125] [,126] [,127] [,128]
#> [1,]    7.7      6    6.9    5.6    7.7    6.3    6.7    7.2    6.2    6.1
#>      [,129] [,130] [,131] [,132] [,133] [,134] [,135] [,136] [,137] [,138]
#> [1,]    6.4    7.2    7.4    7.9    6.4    6.3    6.1    7.7    6.3    6.4
#>      [,139] [,140] [,141] [,142] [,143] [,144] [,145] [,146] [,147] [,148]
#> [1,]      6    6.9    6.7    6.9    5.8    6.8    6.7    6.7    6.3    6.5
#>      [,149] [,150]
#> [1,]    6.2    5.9
#> 
#> $y
#>   [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [38] 1 1 1 1 1 1 1 1 1 1 1 1 1 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#>  [75] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 3 3 3 3 3 3 3 3 3 3 3
#> [112] 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3
#> [149] 3 3
#> 
#> $w
#>   [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [38] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [75] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [112] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [149] 1 1
#> 
#> $l
#> [1] 1 5
#> 
#> $delrsq
#> [1] 0.01
#> 
#> $ns
#> [1] 1
#> 
#> $tx
#>               [,1]
#>   [1,] -0.99612767
#>   [2,] -1.03490874
#>   [3,] -1.18676782
#>   [4,] -1.25337475
#>   [5,] -0.96520326
#>   [6,] -0.49291631
#>   [7,] -1.25337475
#>   [8,] -0.96520326
#>   [9,] -1.32627802
#>  [10,] -1.03490874
#>  [11,] -0.49291631
#>  [12,] -1.11421374
#>  [13,] -1.11421374
#>  [14,] -1.35640999
#>  [15,]  0.21740316
#>  [16,]  0.12248628
#>  [17,] -0.49291631
#>  [18,] -0.99612767
#>  [19,]  0.12248628
#>  [20,] -0.99612767
#>  [21,] -0.49291631
#>  [22,] -0.99612767
#>  [23,] -1.25337475
#>  [24,] -0.99612767
#>  [25,] -1.11421374
#>  [26,] -0.96520326
#>  [27,] -0.96520326
#>  [28,] -0.89523436
#>  [29,] -0.89523436
#>  [30,] -1.18676782
#>  [31,] -1.11421374
#>  [32,] -0.49291631
#>  [33,] -0.89523436
#>  [34,] -0.28236937
#>  [35,] -1.03490874
#>  [36,] -0.96520326
#>  [37,] -0.28236937
#>  [38,] -1.03490874
#>  [39,] -1.32627802
#>  [40,] -0.99612767
#>  [41,] -0.96520326
#>  [42,] -1.29614606
#>  [43,] -1.32627802
#>  [44,] -0.96520326
#>  [45,] -0.99612767
#>  [46,] -1.11421374
#>  [47,] -0.99612767
#>  [48,] -1.25337475
#>  [49,] -0.71575517
#>  [50,] -0.96520326
#>  [51,]  0.90010010
#>  [52,]  0.72099017
#>  [53,]  0.87004482
#>  [54,] -0.28236937
#>  [55,]  0.75355995
#>  [56,]  0.12248628
#>  [57,]  0.65565164
#>  [58,] -1.03490874
#>  [59,]  0.78200310
#>  [60,] -0.89523436
#>  [61,] -0.96520326
#>  [62,]  0.33227956
#>  [63,]  0.45059983
#>  [64,]  0.56008131
#>  [65,] -0.09036787
#>  [66,]  0.80989548
#>  [67,] -0.09036787
#>  [68,]  0.21740316
#>  [69,]  0.61581912
#>  [70,] -0.09036787
#>  [71,]  0.33227956
#>  [72,]  0.56008131
#>  [73,]  0.65565164
#>  [74,]  0.56008131
#>  [75,]  0.72099017
#>  [76,]  0.78200310
#>  [77,]  0.83993012
#>  [78,]  0.80989548
#>  [79,]  0.45059983
#>  [80,]  0.12248628
#>  [81,] -0.28236937
#>  [82,] -0.28236937
#>  [83,]  0.21740316
#>  [84,]  0.45059983
#>  [85,] -0.49291631
#>  [86,]  0.45059983
#>  [87,]  0.80989548
#>  [88,]  0.65565164
#>  [89,] -0.09036787
#>  [90,] -0.28236937
#>  [91,] -0.28236937
#>  [92,]  0.56008131
#>  [93,]  0.21740316
#>  [94,] -0.96520326
#>  [95,] -0.09036787
#>  [96,]  0.12248628
#>  [97,]  0.12248628
#>  [98,]  0.61581912
#>  [99,] -0.99612767
#> [100,]  0.12248628
#> [101,]  0.65565164
#> [102,]  0.21740316
#> [103,]  0.92998009
#> [104,]  0.65565164
#> [105,]  0.75355995
#> [106,]  1.07938007
#> [107,] -1.03490874
#> [108,]  0.98974008
#> [109,]  0.80989548
#> [110,]  0.95986009
#> [111,]  0.75355995
#> [112,]  0.72099017
#> [113,]  0.83993012
#> [114,]  0.12248628
#> [115,]  0.21740316
#> [116,]  0.72099017
#> [117,]  0.75355995
#> [118,]  1.10926007
#> [119,]  1.10926007
#> [120,]  0.45059983
#> [121,]  0.87004482
#> [122,] -0.09036787
#> [123,]  1.10926007
#> [124,]  0.65565164
#> [125,]  0.80989548
#> [126,]  0.95986009
#> [127,]  0.61581912
#> [128,]  0.56008131
#> [129,]  0.72099017
#> [130,]  0.95986009
#> [131,]  1.01962008
#> [132,]  1.16902006
#> [133,]  0.72099017
#> [134,]  0.65565164
#> [135,]  0.56008131
#> [136,]  1.10926007
#> [137,]  0.65565164
#> [138,]  0.72099017
#> [139,]  0.45059983
#> [140,]  0.87004482
#> [141,]  0.80989548
#> [142,]  0.87004482
#> [143,]  0.21740316
#> [144,]  0.83993012
#> [145,]  0.80989548
#> [146,]  0.80989548
#> [147,]  0.65565164
#> [148,]  0.75355995
#> [149,]  0.61581912
#> [150,]  0.33227956
#> 
#> $ty
#>   [1] -1.3531860 -1.3531860 -1.3531860 -1.3531860 -1.3531860 -1.3531860
#>   [7] -1.3531860 -1.3531860 -1.3531860 -1.3531860 -1.3531860 -1.3531860
#>  [13] -1.3531860 -1.3531860 -1.3531860 -1.3531860 -1.3531860 -1.3531860
#>  [19] -1.3531860 -1.3531860 -1.3531860 -1.3531860 -1.3531860 -1.3531860
#>  [25] -1.3531860 -1.3531860 -1.3531860 -1.3531860 -1.3531860 -1.3531860
#>  [31] -1.3531860 -1.3531860 -1.3531860 -1.3531860 -1.3531860 -1.3531860
#>  [37] -1.3531860 -1.3531860 -1.3531860 -1.3531860 -1.3531860 -1.3531860
#>  [43] -1.3531860 -1.3531860 -1.3531860 -1.3531860 -1.3531860 -1.3531860
#>  [49] -1.3531860 -1.3531860  0.3206916  0.3206916  0.3206916  0.3206916
#>  [55]  0.3206916  0.3206916  0.3206916  0.3206916  0.3206916  0.3206916
#>  [61]  0.3206916  0.3206916  0.3206916  0.3206916  0.3206916  0.3206916
#>  [67]  0.3206916  0.3206916  0.3206916  0.3206916  0.3206916  0.3206916
#>  [73]  0.3206916  0.3206916  0.3206916  0.3206916  0.3206916  0.3206916
#>  [79]  0.3206916  0.3206916  0.3206916  0.3206916  0.3206916  0.3206916
#>  [85]  0.3206916  0.3206916  0.3206916  0.3206916  0.3206916  0.3206916
#>  [91]  0.3206916  0.3206916  0.3206916  0.3206916  0.3206916  0.3206916
#>  [97]  0.3206916  0.3206916  0.3206916  0.3206916  1.0324944  1.0324944
#> [103]  1.0324944  1.0324944  1.0324944  1.0324944  1.0324944  1.0324944
#> [109]  1.0324944  1.0324944  1.0324944  1.0324944  1.0324944  1.0324944
#> [115]  1.0324944  1.0324944  1.0324944  1.0324944  1.0324944  1.0324944
#> [121]  1.0324944  1.0324944  1.0324944  1.0324944  1.0324944  1.0324944
#> [127]  1.0324944  1.0324944  1.0324944  1.0324944  1.0324944  1.0324944
#> [133]  1.0324944  1.0324944  1.0324944  1.0324944  1.0324944  1.0324944
#> [139]  1.0324944  1.0324944  1.0324944  1.0324944  1.0324944  1.0324944
#> [145]  1.0324944  1.0324944  1.0324944  1.0324944  1.0324944  1.0324944
#> 
#> $rsq
#> [1] 0.7027773
#> 
#> $ierr
#> [1] 0
#> 
#> $m
#>   [1]  14   9  39  43  42  23  48   7   4  30   3  31  12  13  25  46  35  38
#>  [19] 107   2  58  10  50  94   8   5  27  26  61  44  36  41  40  99  45  47
#>  [37]   1  22  18  24  20  33  29  28  60  49   6  17  21  32  11  85  91  37
#>  [55]  90  54  82  81  34  89 122  70  95  67  65  96  97  19 114  56  16  80
#>  [73] 100  93  15 115  83 143  68 102  71 150  62  63  79 139  84 120  86 135
#>  [91] 128  64  72  92  74  69 149  98 127 124 147  57 104 134 101  88  73 137
#> [109] 133 129 138  52 116  75 112  55 105 117 148 111  76  59 146  66  87  78
#> [127] 141 109 145 125 144  77 113 142 140 121  53  51 103 130 126 110 108 131
#> [145] 106 118 119 123 136 132   1  34  35  36  37  33  32  31  30  29  28  38
#> [163]  39  40  50  49  48  47  46  45  44  43  42  41  27  26  25  11  10   9
#> [181]   8   7   6   5   4   3   2  12  13  14  24  23  22  21  20  19  18  17
#> [199]  16  15  87  86  85  84  83  82  81  80  79  78  77  88  89 100  99  98
#> [217]  97  96  95  94  93  92  91  90  76  75  61  60  59  58  57  56  55  54
#> [235]  53  52  51  62  63  74  73  72  71  70  69  68  67  66  65  64 136 135
#> [253] 134 133 132 131 130 129 128 127 137 138 139 149 148 147 146 145 144 143
#> [271] 142 141 140 126 125 124 110 109 108 107 106 105 104 103 102 101 111 112
#> [289] 113 123 122 121 120 119 118 117 116 115 114 150
#> 
#> $z
#>               [,1]        [,2]       [,3] [,4]         [,5]        [,6]
#>   [1,] -0.99612767 -0.99612767 -0.9077636    1 -0.355757411 -1.35188508
#>   [2,] -0.28236937 -1.03490874 -0.9077636    1 -0.316976338 -1.35188508
#>   [3,] -1.03490874 -1.18676782 -0.9077636    1 -0.165117256 -1.35188508
#>   [4,] -0.96520326 -1.25337475 -0.9077636    1 -0.098510330 -1.35188508
#>   [5,] -0.28236937 -0.96520326 -0.9077636    1 -0.386681821 -1.35188508
#>   [6,] -0.89523436 -0.49291631 -0.9077636    1 -0.858968772 -1.35188508
#>   [7,] -0.49291631 -1.25337475 -0.9077636    1 -0.098510330 -1.35188508
#>   [8,] -1.11421374 -0.96520326 -0.9077636    1 -0.386681821 -1.35188508
#>   [9,] -1.18676782 -1.32627802 -0.9077636    1 -0.025607054 -1.35188508
#>  [10,] -0.89523436 -1.03490874 -0.9077636    1 -0.316976338 -1.35188508
#>  [11,] -0.89523436 -0.49291631 -0.9077636    1 -0.858968772 -1.35188508
#>  [12,] -1.03490874 -1.11421374 -0.9077636    1 -0.237671333 -1.32204236
#>  [13,] -1.32627802 -1.11421374 -0.9077636    1 -0.237671333 -1.32204236
#>  [14,] -0.99612767 -1.35640999 -0.9077636    1  0.004524912 -1.32204236
#>  [15,] -0.96520326  0.21740316 -0.9077636    1 -1.569288234 -1.32204236
#>  [16,] -0.71575517  0.12248628 -0.9077636    1 -1.474371355 -1.32204236
#>  [17,] -1.25337475 -0.49291631 -0.9077636    1 -0.858968772 -0.67881948
#>  [18,] -0.99612767 -0.99612767 -0.9077636    1 -0.355757411 -0.67881948
#>  [19,] -1.11421374  0.12248628 -0.9077636    1 -1.474371355 -0.67881948
#>  [20,] -0.99612767 -0.99612767 -0.9077636    1 -0.355757411 -0.67881948
#>  [21,] -0.96520326 -0.49291631 -0.9077636    1 -0.858968772 -0.67881948
#>  [22,] -1.32627802 -0.99612767 -0.9077636    1 -0.355757411 -0.67881948
#>  [23,] -1.29614606 -1.25337475 -0.9077636    1 -0.098510330 -1.02134897
#>  [24,] -0.96520326 -0.99612767 -0.9077636    1 -0.355757411 -1.02134897
#>  [25,] -0.96520326 -1.11421374 -0.9077636    1 -0.237671333 -1.02134897
#>  [26,] -0.96520326 -0.96520326 -0.9077636    1 -0.386681821 -1.02134897
#>  [27,] -1.11421374 -0.96520326 -0.9077636    1 -0.386681821 -1.02134897
#>  [28,] -0.49291631 -0.89523436 -0.9077636    1 -0.456650716 -1.02134897
#>  [29,] -1.03490874 -0.89523436 -0.9077636    1 -0.456650716 -1.02134897
#>  [30,] -1.32627802 -1.18676782 -0.9077636    1 -0.165117256 -1.02134897
#>  [31,] -0.96520326 -1.11421374 -0.9077636    1 -0.237671333 -1.02134897
#>  [32,] -1.25337475 -0.49291631 -0.9077636    1 -0.858968772 -1.02134897
#>  [33,] -0.49291631 -0.89523436 -0.9077636    1 -0.456650716 -1.19050443
#>  [34,] -0.96520326 -0.28236937 -0.9077636    1 -1.069515712 -1.19050443
#>  [35,] -1.25337475 -1.03490874 -0.9077636    1 -0.316976338 -1.19050443
#>  [36,] -1.18676782 -0.96520326 -0.9077636    1 -0.386681821 -1.19050443
#>  [37,] -1.03490874 -0.28236937 -0.9077636    1 -1.069515712 -1.19050443
#>  [38,] -1.11421374 -1.03490874 -0.9077636    1 -0.316976338 -1.19050443
#>  [39,] -1.11421374 -1.32627802 -0.9077636    1 -0.025607054 -1.19050443
#>  [40,] -1.35640999 -0.99612767 -0.9077636    1 -0.355757411 -1.19050443
#>  [41,] -0.99612767 -0.96520326 -0.9077636    1 -0.386681821 -1.19050443
#>  [42,] -1.25337475 -1.29614606 -0.9077636    1 -0.055739021 -1.13177041
#>  [43,] -0.99612767 -1.32627802 -0.9077636    1 -0.025607054 -1.13177041
#>  [44,] -0.49291631 -0.96520326 -0.9077636    1 -0.386681821 -1.13177041
#>  [45,] -0.99612767 -0.99612767 -0.9077636    1 -0.355757411 -1.13177041
#>  [46,]  0.12248628 -1.11421374 -0.9077636    1 -0.237671333 -1.16652536
#>  [47,] -0.99612767 -0.99612767 -0.9077636    1 -0.355757411 -1.07186429
#>  [48,] -0.49291631 -1.25337475 -0.9077636    1 -0.098510330 -1.07186429
#>  [49,]  0.12248628 -0.71575517 -0.9077636    1 -0.636129907 -1.07186429
#>  [50,]  0.21740316 -0.96520326 -0.9077636    1 -0.386681821 -1.07186429
#>  [51,]  0.80989548  0.90010010  0.2151309    1 -0.583747675 -1.07186429
#>  [52,]  0.45059983  0.72099017  0.2151309    1 -0.404637754 -1.07186429
#>  [53,] -0.49291631  0.87004482  0.2151309    1 -0.553692404 -0.08727783
#>  [54,]  0.45059983 -0.28236937  0.2151309    1  0.598721786 -0.08727783
#>  [55,]  0.21740316  0.75355995  0.2151309    1 -0.437207525 -0.08727783
#>  [56,] -0.28236937  0.12248628  0.2151309    1  0.193866143 -0.08727783
#>  [57,] -0.28236937  0.65565164  0.2151309    1 -0.339299220 -0.08727783
#>  [58,]  0.12248628 -1.03490874  0.2151309    1  1.351261160 -0.08727783
#>  [59,]  0.45059983  0.78200310  0.2151309    1 -0.465650677 -0.08727783
#>  [60,]  0.80989548 -0.89523436  0.2151309    1  1.211586782  0.36775002
#>  [61,]  0.83993012 -0.96520326  0.2151309    1  1.281555677  0.36775002
#>  [62,]  0.65565164  0.33227956  0.2151309    1 -0.015927139  0.36775002
#>  [63,] -0.09036787  0.45059983  0.2151309    1 -0.134247405  0.36775002
#>  [64,]  0.12248628  0.56008131  0.2151309    1 -0.243728890  0.36775002
#>  [65,] -0.99612767 -0.09036787  0.2151309    1  0.406720289  0.36775002
#>  [66,]  0.61581912  0.80989548  0.2151309    1 -0.493543057 -0.01541966
#>  [67,]  0.12248628 -0.09036787  0.2151309    1  0.406720289 -0.01541966
#>  [68,]  0.12248628  0.21740316  0.2151309    1  0.098949264 -0.01541966
#>  [69,] -0.09036787  0.61581912  0.2151309    1 -0.299466697 -0.01541966
#>  [70,] -0.96520326 -0.09036787  0.2151309    1  0.406720289 -0.01541966
#>  [71,]  0.21740316  0.33227956  0.2151309    1 -0.015927139 -0.01541966
#>  [72,]  0.56008131  0.56008131  0.2151309    1 -0.243728890 -0.01541966
#>  [73,] -0.28236937  0.65565164  0.2151309    1 -0.339299220 -0.01541966
#>  [74,] -0.28236937  0.56008131  0.2151309    1 -0.243728890  0.40588435
#>  [75,]  0.78200310  0.72099017  0.2151309    1 -0.404637754  0.40588435
#>  [76,]  0.72099017  0.78200310  0.2151309    1 -0.465650677  0.40588435
#>  [77,] -0.96520326  0.83993012  0.2151309    1 -0.523577695  0.40588435
#>  [78,] -0.89523436  0.80989548  0.2151309    1 -0.493543057  0.40588435
#>  [79,]  0.78200310  0.45059983  0.2151309    1 -0.134247405  0.40588435
#>  [80,] -1.03490874  0.12248628  0.2151309    1  0.193866143  0.40588435
#>  [81,]  0.65565164 -0.28236937  0.2151309    1  0.598721786  0.57632276
#>  [82,]  0.12248628 -0.28236937  0.2151309    1  0.598721786  0.57632276
#>  [83,]  0.75355995  0.21740316  0.2151309    1  0.098949264  0.57632276
#>  [84,] -0.28236937  0.45059983  0.2151309    1 -0.134247405  0.58663543
#>  [85,]  0.87004482 -0.49291631  0.2151309    1  0.809268726  0.58663543
#>  [86,]  0.72099017  0.45059983  0.2151309    1 -0.134247405  0.58663543
#>  [87,]  0.90010010  0.80989548  0.2151309    1 -0.493543057  0.58663543
#>  [88,]  0.33227956  0.65565164  0.2151309    1 -0.339299220  0.58663543
#>  [89,]  0.45059983 -0.09036787  0.2151309    1  0.406720289  0.58663543
#>  [90,]  0.56008131 -0.28236937  0.2151309    1  0.598721786  0.54194143
#>  [91,]  0.65565164 -0.28236937  0.2151309    1  0.598721786  0.54194143
#>  [92,]  0.56008131  0.56008131  0.2151309    1 -0.243728890  0.54194143
#>  [93,]  0.33227956  0.21740316  0.2151309    1  0.098949264  0.54194143
#>  [94,] -0.09036787 -0.96520326  0.2151309    1  1.281555677  0.54194143
#>  [95,]  0.61581912 -0.09036787  0.2151309    1  0.406720289  0.54194143
#>  [96,]  0.21740316  0.12248628  0.2151309    1  0.193866143  0.64960892
#>  [97,] -0.09036787  0.12248628  0.2151309    1  0.193866143  0.64960892
#>  [98,]  0.80989548  0.61581912  0.2151309    1 -0.299466697  0.64960892
#>  [99,] -0.09036787 -0.99612767  0.2151309    1  1.312480087  0.64960892
#> [100,]  0.56008131  0.12248628  0.2151309    1  0.193866143  0.80398706
#> [101,]  1.10926007  0.65565164  0.6926327    1  0.379881016  0.80398706
#> [102,]  0.56008131  0.21740316  0.6926327    1  0.818129500  0.80398706
#> [103,]  0.65565164  0.92998009  0.6926327    1  0.105552565  0.80398706
#> [104,]  0.72099017  0.65565164  0.6926327    1  0.379881016  0.80398706
#> [105,]  1.16902006  0.75355995  0.6926327    1  0.281972711  0.80398706
#> [106,]  1.01962008  1.07938007  0.6926327    1 -0.043847417  0.80398706
#> [107,]  0.95986009 -1.03490874  0.6926327    1  2.070441396  0.80398706
#> [108,]  0.72099017  0.98974008  0.6926327    1  0.045792572  0.80398706
#> [109,]  0.56008131  0.80989548  0.6926327    1  0.225637180  0.81244001
#> [110,]  0.61581912  0.95986009  0.6926327    1  0.075672569  0.81244001
#> [111,]  0.65565164  0.75355995  0.6926327    1  0.281972711  0.81244001
#> [112,]  0.72099017  0.72099017  0.6926327    1  0.314542482  0.81244001
#> [113,]  0.45059983  0.83993012  0.6926327    1  0.195602541  0.81244001
#> [114,]  0.61581912  0.12248628  0.6926327    1  0.913046380  0.81244001
#> [115,]  0.75355995  0.21740316  0.6926327    1  0.818129500  0.81244001
#> [116,]  0.65565164  0.72099017  0.6926327    1  0.314542482  0.80632725
#> [117,]  0.80989548  0.75355995  0.6926327    1  0.281972711  0.80632725
#> [118,]  0.80989548  1.10926007  0.6926327    1 -0.073727413  0.80632725
#> [119,]  0.83993012  1.10926007  0.6926327    1 -0.073727413  0.80632725
#> [120,]  0.21740316  0.45059983  0.6926327    1  0.584932832  0.80632725
#> [121,]  0.87004482  0.87004482  0.6926327    1  0.165487832  0.67594254
#> [122,]  0.80989548 -0.09036787  0.6926327    1  1.125900526  0.67594254
#> [123,]  0.87004482  1.10926007  0.6926327    1 -0.073727413  0.72448996
#> [124,]  0.95986009  0.65565164  0.6926327    1  0.379881016  0.72448996
#> [125,]  0.80989548  0.80989548  0.6926327    1  0.225637180  0.72448996
#> [126,]  0.65565164  0.95986009  0.6926327    1  0.075672569  0.72448996
#> [127,]  0.95986009  0.61581912  0.6926327    1  0.419713539  0.72448996
#> [128,]  0.80989548  0.56008131  0.6926327    1  0.475451347  0.72448996
#> [129,]  0.98974008  0.72099017  0.6926327    1  0.314542482  0.72448996
#> [130,] -1.03490874  0.95986009  0.6926327    1  0.075672569  0.72448996
#> [131,]  1.07938007  1.01962008  0.6926327    1  0.015912576  0.93218379
#> [132,]  0.75355995  1.16902006  0.6926327    1 -0.133487406  0.93218379
#> [133,]  0.65565164  0.72099017  0.6926327    1  0.314542482  0.93218379
#> [134,]  0.92998009  0.65565164  0.6926327    1  0.379881016  0.78380831
#> [135,]  0.21740316  0.56008131  0.6926327    1  0.475451347  0.78380831
#> [136,]  0.65565164  1.10926007  0.6926327    1 -0.073727413  0.78380831
#> [137,]  0.75355995  0.65565164  0.6926327    1  0.379881016  0.78380831
#> [138,]  0.72099017  0.72099017  0.6926327    1  0.314542482  0.85074329
#> [139,]  0.83993012  0.45059983  0.6926327    1  0.584932832  0.89660011
#> [140,]  1.10926007  0.87004482  0.6926327    1  0.165487832  0.90565744
#> [141,] -0.09036787  0.80989548  0.6926327    1  0.225637180  0.90565744
#> [142,]  0.87004482  0.87004482  0.6926327    1  0.165487832  0.90565744
#> [143,]  0.45059983  0.21740316  0.6926327    1  0.818129500  1.03553266
#> [144,]  1.10926007  0.83993012  0.6926327    1  0.195602541  1.03553266
#> [145,]  1.10926007  0.80989548  0.6926327    1  0.225637180  1.03553266
#> [146,]  0.75355995  0.80989548  0.6926327    1  0.225637180  1.03553266
#> [147,]  0.72099017  0.65565164  0.6926327    1  0.379881016  1.03553266
#> [148,]  0.21740316  0.75355995  0.6926327    1  0.281972711  1.03553266
#> [149,]  0.12248628  0.61581912  0.6926327    1  0.419713539  1.03553266
#> [150,]  0.33227956  0.33227956  0.6926327    1  0.703253097  1.03553266
#>             [,7]        [,8]       [,9]       [,10]       [,11]        [,12]
#>   [1,] 0.0500000 -1.52839529 -1.3518851 -1.80787338  0.08488197   0.05000000
#>   [2,] 0.0500000 -1.44206008 -1.3518851 -1.68468312  0.15581024   0.05000000
#>   [3,] 0.0500000 -1.44206008 -1.3518851 -1.68468312  0.15581024   0.05000000
#>   [4,] 0.0500000 -1.44206008 -1.3518851 -1.68468312  0.15581024   0.05000000
#>   [5,] 0.1158917 -1.35572486 -1.3535718 -1.56149286  0.22673852   0.05000000
#>   [6,] 0.1843083 -1.26938965 -1.2780196 -1.43830261  0.29766679   0.06334044
#>   [7,] 0.1843083 -1.26938965 -1.2780196 -1.43830261  0.29766679   0.06334044
#>   [8,] 0.1843083 -1.26938965 -1.2780196 -1.43830261  0.29766679   0.06334044
#>   [9,] 0.1843083 -1.26938965 -1.2780196 -1.43830261  0.29766679   0.06334044
#>  [10,] 0.2527249 -1.18305444 -1.2062636 -1.31511235  0.36859507   0.21842166
#>  [11,] 0.2527249 -1.18305444 -1.2062636 -1.31511235  0.36859507   0.21842166
#>  [12,] 0.3211416 -1.09671922 -1.1351626 -1.19192210  0.43952334   0.44774905
#>  [13,] 0.3211416 -1.09671922 -1.1351626 -1.19192210  0.43952334   0.44774905
#>  [14,] 0.3211416 -1.09671922 -1.1351626 -1.19192210  0.43952334   0.44774905
#>  [15,] 0.3211416 -1.09671922 -1.1351626 -1.19192210  0.43952334   0.44774905
#>  [16,] 0.3211416 -1.09671922 -1.1351626 -1.19192210  0.43952334   0.44774905
#>  [17,] 0.3565608 -1.03800622 -1.0540410 -1.06873184  0.49678327   0.50000000
#>  [18,] 0.3565608 -1.03800622 -1.0540410 -1.06873184  0.49678327   0.50000000
#>  [19,] 0.3565608 -1.03800622 -1.0540410 -1.06873184  0.49678327   0.50000000
#>  [20,] 0.3565608 -1.03800622 -1.0540410 -1.06873184  0.49678327   0.50000000
#>  [21,] 0.3565608 -1.03800622 -1.0540410 -1.06873184  0.49678327   0.50000000
#>  [22,] 0.3565608 -1.03800622 -1.0540410 -1.06873184  0.49678327   0.50000000
#>  [23,] 0.3436970 -1.02118505 -0.9849526 -0.94554159  0.60239754   0.37399312
#>  [24,] 0.3436970 -1.02118505 -0.9849526 -0.94554159  0.60239754   0.37399312
#>  [25,] 0.3436970 -1.02118505 -0.9849526 -0.94554159  0.60239754   0.37399312
#>  [26,] 0.3436970 -1.02118505 -0.9849526 -0.94554159  0.60239754   0.37399312
#>  [27,] 0.3436970 -1.02118505 -0.9849526 -0.94554159  0.60239754   0.37399312
#>  [28,] 0.3436970 -1.02118505 -0.9849526 -0.94554159  0.60239754   0.37399312
#>  [29,] 0.3436970 -1.02118505 -0.9849526 -0.94554159  0.60239754   0.37399312
#>  [30,] 0.3436970 -1.02118505 -0.9849526 -0.94554159  0.60239754   0.37399312
#>  [31,] 0.3436970 -1.02118505 -0.9849526 -0.94554159  0.60239754   0.37399312
#>  [32,] 0.3436970 -1.02118505 -0.9849526 -0.94554159  0.60239754   0.37399312
#>  [33,] 0.2651767 -1.06931755 -1.0158770 -0.82333754  0.66124572   0.19899248
#>  [34,] 0.2651767 -1.06931755 -1.0158770 -0.82333754  0.66124572   0.19899248
#>  [35,] 0.2651767 -1.06931755 -1.0158770 -0.82333754  0.66124572   0.19899248
#>  [36,] 0.2651767 -1.06931755 -1.0158770 -0.82333754  0.66124572   0.19899248
#>  [37,] 0.2651767 -1.06931755 -1.0158770 -0.82333754  0.66124572   0.19899248
#>  [38,] 0.2651767 -1.06931755 -1.0158770 -0.82333754  0.66124572   0.19899248
#>  [39,] 0.2651767 -1.06931755 -1.0158770 -0.82333754  0.66124572   0.19899248
#>  [40,] 0.2651767 -1.06931755 -1.0158770 -0.82333754  0.66124572   0.19899248
#>  [41,] 0.2651767 -1.06931755 -1.0158770 -0.82333754  0.66124572   0.19899248
#>  [42,] 0.2144485 -0.96416432 -0.9513637 -0.69837931  0.68323186   0.14791147
#>  [43,] 0.2144485 -0.96416432 -0.9513637 -0.69837931  0.68323186   0.14791147
#>  [44,] 0.2144485 -0.96416432 -0.9513637 -0.69837931  0.68323186   0.14791147
#>  [45,] 0.2144485 -0.96416432 -0.9513637 -0.69837931  0.68323186   0.14791147
#>  [46,] 0.1982661 -0.77490422 -0.7794311 -0.55595565  0.72597347   0.15034102
#>  [47,] 0.2087562 -0.51181428 -0.5086637 -0.40387003  0.72127990   0.12950576
#>  [48,] 0.2087562 -0.51181428 -0.5086637 -0.40387003  0.72127990   0.12950576
#>  [49,] 0.2087562 -0.51181428 -0.5086637 -0.40387003  0.72127990   0.12950576
#>  [50,] 0.2087562 -0.51181428 -0.5086637 -0.40387003  0.72127990   0.12950576
#>  [51,] 0.2087562 -0.51181428 -0.5086637 -0.40387003  0.72127990   0.12950576
#>  [52,] 0.2087562 -0.51181428 -0.5086637 -0.40387003  0.72127990   0.12950576
#>  [53,] 0.2459444 -0.30970893 -0.3018605 -0.25846160  0.73757831   0.20614326
#>  [54,] 0.2459444 -0.30970893 -0.3018605 -0.25846160  0.73757831   0.20614326
#>  [55,] 0.2459444 -0.30970893 -0.3018605 -0.25846160  0.73757831   0.20614326
#>  [56,] 0.2459444 -0.30970893 -0.3018605 -0.25846160  0.73757831   0.20614326
#>  [57,] 0.2459444 -0.30970893 -0.3018605 -0.25846160  0.73757831   0.20614326
#>  [58,] 0.2459444 -0.30970893 -0.3018605 -0.25846160  0.73757831   0.20614326
#>  [59,] 0.2459444 -0.30970893 -0.3018605 -0.25846160  0.73757831   0.20614326
#>  [60,] 0.2883603 -0.09253349 -0.1113235 -0.15632928  0.72126316   0.32629084
#>  [61,] 0.2883603 -0.09253349 -0.1113235 -0.15632928  0.72126316   0.32629084
#>  [62,] 0.2883603 -0.09253349 -0.1113235 -0.15632928  0.72126316   0.32629084
#>  [63,] 0.2883603 -0.09253349 -0.1113235 -0.15632928  0.72126316   0.32629084
#>  [64,] 0.2883603 -0.09253349 -0.1113235 -0.15632928  0.72126316   0.32629084
#>  [65,] 0.2883603 -0.09253349 -0.1113235 -0.15632928  0.72126316   0.32629084
#>  [66,] 0.3437503  0.21587446  0.1027369 -0.02023811  0.62617684   0.38458092
#>  [67,] 0.3437503  0.21587446  0.1027369 -0.02023811  0.62617684   0.38458092
#>  [68,] 0.3437503  0.21587446  0.1027369 -0.02023811  0.62617684   0.38458092
#>  [69,] 0.3437503  0.21587446  0.1027369 -0.02023811  0.62617684   0.38458092
#>  [70,] 0.3437503  0.21587446  0.1027369 -0.02023811  0.62617684   0.38458092
#>  [71,] 0.3437503  0.21587446  0.1027369 -0.02023811  0.62617684   0.38458092
#>  [72,] 0.3437503  0.21587446  0.1027369 -0.02023811  0.62617684   0.38458092
#>  [73,] 0.3437503  0.21587446  0.1027369 -0.02023811  0.62617684   0.38458092
#>  [74,] 0.3865413  0.32450776  0.1975222  0.12028675  0.54685119   0.35961228
#>  [75,] 0.3865413  0.32450776  0.1975222  0.12028675  0.54685119   0.35961228
#>  [76,] 0.3865413  0.32450776  0.1975222  0.12028675  0.54685119   0.35961228
#>  [77,] 0.3865413  0.32450776  0.1975222  0.12028675  0.54685119   0.35961228
#>  [78,] 0.3865413  0.32450776  0.1975222  0.12028675  0.54685119   0.35961228
#>  [79,] 0.3865413  0.32450776  0.1975222  0.12028675  0.54685119   0.35961228
#>  [80,] 0.3865413  0.32450776  0.1975222  0.12028675  0.54685119   0.35961228
#>  [81,] 0.4133837  0.36674214  0.3087894  0.28526538  0.49643121   0.42363020
#>  [82,] 0.4133837  0.36674214  0.3087894  0.28526538  0.49643121   0.42363020
#>  [83,] 0.4133837  0.36674214  0.3087894  0.28526538  0.49643121   0.42363020
#>  [84,] 0.4341305  0.53110802  0.4315865  0.40358743  0.41471282   0.43610155
#>  [85,] 0.4341305  0.53110802  0.4315865  0.40358743  0.41471282   0.43610155
#>  [86,] 0.4341305  0.53110802  0.4315865  0.40358743  0.41471282   0.43610155
#>  [87,] 0.4341305  0.53110802  0.4315865  0.40358743  0.41471282   0.43610155
#>  [88,] 0.4341305  0.53110802  0.4315865  0.40358743  0.41471282   0.43610155
#>  [89,] 0.4341305  0.53110802  0.4315865  0.40358743  0.41471282   0.43610155
#>  [90,] 0.4552605  0.65776757  0.5429901  0.52287303  0.36368218   0.46597685
#>  [91,] 0.4552605  0.65776757  0.5429901  0.52287303  0.36368218   0.46597685
#>  [92,] 0.4552605  0.65776757  0.5429901  0.52287303  0.36368218   0.46597685
#>  [93,] 0.4552605  0.65776757  0.5429901  0.52287303  0.36368218   0.46597685
#>  [94,] 0.4552605  0.65776757  0.5429901  0.52287303  0.36368218   0.46597685
#>  [95,] 0.4552605  0.65776757  0.5429901  0.52287303  0.36368218   0.46597685
#>  [96,] 0.4660413  0.69977770  0.6025609  0.59015174  0.34128537   0.50000000
#>  [97,] 0.4660413  0.69977770  0.6025609  0.59015174  0.34128537   0.50000000
#>  [98,] 0.4660413  0.69977770  0.6025609  0.59015174  0.34128537   0.50000000
#>  [99,] 0.4660413  0.69977770  0.6025609  0.59015174  0.34128537   0.50000000
#> [100,] 0.4703518  0.73411124  0.6359023  0.62513220  0.35425349   0.48691977
#> [101,] 0.4703518  0.73411124  0.6359023  0.62513220  0.35425349   0.48691977
#> [102,] 0.4703518  0.73411124  0.6359023  0.62513220  0.35425349   0.48691977
#> [103,] 0.4703518  0.73411124  0.6359023  0.62513220  0.35425349   0.48691977
#> [104,] 0.4703518  0.73411124  0.6359023  0.62513220  0.35425349   0.48691977
#> [105,] 0.4703518  0.73411124  0.6359023  0.62513220  0.35425349   0.48691977
#> [106,] 0.4703518  0.73411124  0.6359023  0.62513220  0.35425349   0.48691977
#> [107,] 0.4703518  0.73411124  0.6359023  0.62513220  0.35425349   0.48691977
#> [108,] 0.4703518  0.73411124  0.6359023  0.62513220  0.35425349   0.48691977
#> [109,] 0.4757503  0.76560629  0.7017687  0.69615473  0.37030694   0.43837168
#> [110,] 0.4757503  0.76560629  0.7017687  0.69615473  0.37030694   0.43837168
#> [111,] 0.4757503  0.76560629  0.7017687  0.69615473  0.37030694   0.43837168
#> [112,] 0.4757503  0.76560629  0.7017687  0.69615473  0.37030694   0.43837168
#> [113,] 0.4757503  0.76560629  0.7017687  0.69615473  0.37030694   0.43837168
#> [114,] 0.4757503  0.76560629  0.7017687  0.69615473  0.37030694   0.43837168
#> [115,] 0.4757503  0.76560629  0.7017687  0.69615473  0.37030694   0.43837168
#> [116,] 0.4765942  0.77494070  0.7343828  0.73095078  0.35489215   0.47205745
#> [117,] 0.4765942  0.77494070  0.7343828  0.73095078  0.35489215   0.47205745
#> [118,] 0.4765942  0.77494070  0.7343828  0.73095078  0.35489215   0.47205745
#> [119,] 0.4765942  0.77494070  0.7343828  0.73095078  0.35489215   0.47205745
#> [120,] 0.4765942  0.77494070  0.7343828  0.73095078  0.35489215   0.47205745
#> [121,] 0.4811323  0.78260915  0.7622005  0.76083078  0.33539482   0.50000000
#> [122,] 0.4811323  0.78260915  0.7622005  0.76083078  0.33539482   0.50000000
#> [123,] 0.4866176  0.77826900  0.7901558  0.79071077  0.32803592   0.50000000
#> [124,] 0.4866176  0.77826900  0.7901558  0.79071077  0.32803592   0.50000000
#> [125,] 0.4866176  0.77826900  0.7901558  0.79071077  0.32803592   0.50000000
#> [126,] 0.4866176  0.77826900  0.7901558  0.79071077  0.32803592   0.50000000
#> [127,] 0.4866176  0.77826900  0.7901558  0.79071077  0.32803592   0.50000000
#> [128,] 0.4866176  0.77826900  0.7901558  0.79071077  0.32803592   0.50000000
#> [129,] 0.4866176  0.77826900  0.7901558  0.79071077  0.32803592   0.50000000
#> [130,] 0.4866176  0.77826900  0.7901558  0.79071077  0.32803592   0.50000000
#> [131,] 0.4771298  0.81200095  0.8199359  0.82059077  0.29830513   0.50000000
#> [132,] 0.4771298  0.81200095  0.8199359  0.82059077  0.29830513   0.50000000
#> [133,] 0.4771298  0.81200095  0.8199359  0.82059077  0.29830513   0.50000000
#> [134,] 0.5000000  0.81048963  0.8504708  0.85047076  0.27824281   0.50000000
#> [135,] 0.5000000  0.81048963  0.8504708  0.85047076  0.27824281   0.50000000
#> [136,] 0.5000000  0.81048963  0.8504708  0.85047076  0.27824281   0.50000000
#> [137,] 0.5000000  0.81048963  0.8504708  0.85047076  0.27824281   0.50000000
#> [138,] 0.5000000  0.83930699  0.8803508  0.88035076  0.25004618   0.50000000
#> [139,] 0.5000000  0.87226236  0.9102308  0.91023076  0.22040144   0.50000000
#> [140,] 0.5000000  0.90521773  0.9401108  0.94011075  0.19075670   0.50000000
#> [141,] 0.5000000  0.90521773  0.9401108  0.94011075  0.19075670   0.50000000
#> [142,] 0.5000000  0.90521773  0.9401108  0.94011075  0.19075670   0.50000000
#> [143,] 0.5000000  0.93817309  0.9699908  0.96999075  0.16111196   0.40458555
#> [144,] 0.5000000  0.97112846  0.9998707  0.99987075  0.13146721   0.28069744
#> [145,] 0.5000000  1.03703919  1.0596307  1.05963074  0.07217773   0.06782418
#> [146,] 0.5000000  1.06999455  1.0895107  1.08951074  0.04253299   0.05000000
#> [147,] 0.5000000  1.06999455  1.0895107  1.08951074  0.04253299   0.05000000
#> [148,] 0.5000000  1.06999455  1.0895107  1.08951074  0.04253299   0.05000000
#> [149,] 0.5000000  1.06999455  1.0895107  1.08951074  0.04253299   0.05000000
#> [150,] 0.5000000  1.13590528  1.1492707  1.14927073 -0.01675649 979.23190903
#>