bullseye: Visualising Multiple Pairwise Variable Correlations and Other Scores

We provide a tidy data structure and visualisations for multiple or grouped variable correlations, general association measures scagnostics and other pairwise scores suitable for numerical, ordinal and nominal variables. Supported measures include distance correlation, maximal information, ace correlation, Kendall's tau, and polychoric corrrelation.

See also

Author

Maintainer: Catherine Hurley catherine.hurley@mu.ie

Authors:

Amit Chinwan amit.chinwan.2019@mumail.ie