Graph-theoretic scagnostics values

Calculates scagnostic values for every numeric variable pair in a dataset.

Usage

pair_scagnostics(
  d,
  scagnostic = c("Outlying", "Skewed", "Clumpy", "Sparse", "Striated", "Convex",
    "Skinny", "Stringy", "Monotonic"),
  handle.na = T,
  ...
)

Arguments

d

A dataframe

scagnostic

a character vector for the scagnostic to be calculated. Subset of "Outlying", "Stringy", "Striated", "Clumpy", "Sparse", "Skewed", "Convex", "Skinny" or "Monotonic"

handle.na

If TRUE uses pairwise complete observations.

...

other arguments

Value

A tibble of class pairwise with scagnostic values for every numeric variable pair, or NULL if there are not at least two numeric variables

Details

The scagnostic values are calculated using scagnostics function from the scagnostics package.

References

Wilkinson, Leland, Anushka Anand, and Robert Grossman. "Graph-theoretic scagnostics." Information Visualization, IEEE Symposium on. IEEE Computer Society, 2005

Examples

 pair_scagnostics(iris)
#> # A tibble: 54 × 6
#>    x            y            score    group  value pair_type
#>    <chr>        <chr>        <chr>    <chr>  <dbl> <chr>    
#>  1 Petal.Length Sepal.Length Outlying all   0.247  nn       
#>  2 Petal.Width  Sepal.Length Outlying all   0.0767 nn       
#>  3 Sepal.Length Sepal.Width  Outlying all   0.168  nn       
#>  4 Petal.Length Sepal.Width  Outlying all   0.263  nn       
#>  5 Petal.Width  Sepal.Width  Outlying all   0.451  nn       
#>  6 Petal.Length Petal.Width  Outlying all   0.0484 nn       
#>  7 Petal.Length Sepal.Length Skewed   all   0.643  nn       
#>  8 Petal.Width  Sepal.Length Skewed   all   0.743  nn       
#>  9 Sepal.Length Sepal.Width  Skewed   all   0.736  nn       
#> 10 Petal.Length Sepal.Width  Skewed   all   0.578  nn       
#> # ℹ 44 more rows