Constructs a pairwise result for each level of a by variable.

Constructs a pairwise result for each level of a by variable.

Usage

pairwise_by(d, by, pair_fun, ungrouped = TRUE)

Arguments

d

a dataframe

by

a character string for the name of the conditioning variable.

pair_fun

One of the pair_ functions

ungrouped

If TRUE calculates the ungrouped score in addition to grouped scores.

Value

tibble of class "pairwise"

Examples

pairwise_by(iris, by="Species", pair_cor)
#> # A tibble: 24 × 6
#>    x            y            score   group      value pair_type
#>    <chr>        <chr>        <chr>   <fct>      <dbl> <chr>    
#>  1 Petal.Length Sepal.Length pearson setosa     0.267 nn       
#>  2 Petal.Width  Sepal.Length pearson setosa     0.278 nn       
#>  3 Sepal.Length Sepal.Width  pearson setosa     0.743 nn       
#>  4 Petal.Length Sepal.Width  pearson setosa     0.178 nn       
#>  5 Petal.Width  Sepal.Width  pearson setosa     0.233 nn       
#>  6 Petal.Length Petal.Width  pearson setosa     0.332 nn       
#>  7 Petal.Length Sepal.Length pearson versicolor 0.754 nn       
#>  8 Petal.Width  Sepal.Length pearson versicolor 0.546 nn       
#>  9 Sepal.Length Sepal.Width  pearson versicolor 0.526 nn       
#> 10 Petal.Length Sepal.Width  pearson versicolor 0.561 nn       
#> # ℹ 14 more rows